backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
requests
gunicorn
